requirements.txt
gitpython
transformers>=4.0.0
huggingface_hub>=0.0.17
PyYAML>=5.4.1
